Construct help and completion text optimally
For shell completion text, the padding is the longest occurring command
name subtracted, by the length of the current command.

For the `help` subcommand text, the padding is the longest occurring
short-description of a command, subtracted by the length of the current
command's short-description.

Optimally here means with the minimum whitespace between sections of the
text. (By section, we mean either a command name or a command
short-description.)

For example (short completion text):

  inspect-local-release  - Display information from release metadata
  inspect-local-stemcell - Display information from stemcell metadata
  inspect-release        - List release contents such as jobs

`inspect-local-stemcell` is the longest section (22 characters) and its
length is used as padding between the other entries, i.e.
`inspect-local-release` has one whitespace padding between it and the
`-`.